+++ simconvoi.h
@@ -435,9 +435,9 @@ private:
 	bool reversed;
 
 	uint32 heaviest_vehicle;
-	uint16 longest_min_loading_time;
-	uint16 longest_max_loading_time;
-	uint16 current_loading_time;
+	uint32 longest_min_loading_time;
+	uint32 longest_max_loading_time;
+	uint32 current_loading_time;
 
 	/**
 	 * Time in ticks since this convoy last departed from
@@ -1068,10 +1068,10 @@ public:
 	inline uint32 get_heaviest_vehicle() const { return heaviest_vehicle; }
 	
 	//@author: jamespetts
-	uint16 calc_longest_min_loading_time();
-	uint16 calc_longest_max_loading_time();
-	inline uint16 get_longest_min_loading_time() const { return longest_min_loading_time; }
-	inline uint16 get_longest_max_loading_time() const { return longest_max_loading_time; }
+	uint32 calc_longest_min_loading_time();
+	uint32 calc_longest_max_loading_time();
+	inline uint32 get_longest_min_loading_time() const { return longest_min_loading_time; }
+	inline uint32 get_longest_max_loading_time() const { return longest_max_loading_time; }
 
 	void calc_current_loading_time(uint16 load_charge);
 	inline uint16 get_current_loading_time() const { return current_loading_time; }
